modify user package structure
