--- domain/model/auth/auth.go
@@ -1,16 +1,0 @@
-package auth
-
-import "github.com/the-gleam/kiwi-basket/domain/model/user"
-
-type Auth struct {
-	username user.Username
-	token    string
-}
-
-func (a Auth) Token() string {
-	return a.token
-}
-
-func (a Auth) Username() user.Username {
-	return a.username
-}
