+++ domain/model/user/username/username.go
@@ -1,4 +1,4 @@
-package user
+package username
 
 import "fmt"
