+++ domain/model/user/username/username_test.go
@@ -1,4 +1,4 @@
-package user
+package username
 
 import (
 	"testing"
